refactor: added dedicated `protocol` module
